+++ ui/repl.c
@@ -87,7 +87,8 @@ static const char *opts =
     " --check-bounds={yes|no}  Emit bounds checks always or never (ignoring declarations)\n"
     " -O, --optimize           Run time-intensive code optimizations\n"
     " --int-literals={32|64}   Select integer literal size independent of platform\n"
-    " --dump-bitcode={yes|no}  Dump bitcode for the system image (used with --build)\n";
+    " --dump-bitcode={yes|no}  Dump bitcode for the system image (used with --build)\n"
+    " --no-depwarn             Turn off syntax and method deprecation warnings\n";
 
 void parse_opts(int *argcp, char ***argvp)
 {
@@ -106,6 +107,7 @@ void parse_opts(int *argcp, char ***argvp)
         { "int-literals",  required_argument, 0, 301 },
         { "dump-bitcode",  required_argument, 0, 302 },
         { "compile",       required_argument, 0, 303 },
+        { "no-depwarn",    no_argument,       0, 304 },
         { 0, 0, 0, 0 }
     };
     int c;
@@ -198,6 +200,9 @@ void parse_opts(int *argcp, char ***argvp)
                 exit(1);
             }
             break;
+	case 304:
+	    jl_compileropts.depwarn = 0;
+	    break;
         default:
             ios_printf(ios_stderr, "julia: unhandled option -- %c\n",  c);
             ios_printf(ios_stderr, "This is a bug, please report it.\n");